member controller name fix
