+++ src/controllers/member-controller.js
@@ -16,6 +16,20 @@ const updateMember = async (req, res, next) => {
     }
 }
 
+const getListMember = async (req,res)=>{
+    try {
+    const { query } = req;
+        
+     const members= await memberService.getListAllMembers(query)
+
+     res.status(200).json(members) 
+    } catch (err) {
+        res.status(500).json(err.message)
+        console.log(err)
+    }
+}
+
 module.exports ={
-    updateMember
+    updateMember,
+    getListMember
 }
